pdf export implemented, pop up for buttons implemented

+++ SWE2TourPlanner/ViewModels/MainViewModel.cs
@@ -20,6 +20,11 @@ namespace SWE2TourPlanner.ViewModels
         private TourItem currentItem;
         private LogItem currentLog;
         private string searchText;
+        private bool isPopUpDeleteTourVisible;
+        private bool isPopUpModifyTourVisible;
+        private bool isPopUpDeleteLogVisible;
+        private bool isPopUpModifyLogVisible;
+        private bool isPopUpAddLogVisible;
 
         private ICommand popUpAddCommand;
         private ICommand deleteTourCommand;
@@ -29,6 +34,17 @@ namespace SWE2TourPlanner.ViewModels
         private ICommand popUpModifyLogCommand;
         private ICommand searchCommand;
         private ICommand clearCommand;
+        private ICommand openDeleteTourPopUpCommand;
+        private ICommand closeDeleteTourPopUpCommand;
+        private ICommand openModifyTourPopUpCommand;
+        private ICommand closeModifyTourPopUpCommand;
+        private ICommand openDeleteLogPopUpCommand;
+        private ICommand closeDeleteLogPopUpCommand;
+        private ICommand openModifyLogPopUpCommand;
+        private ICommand closeModifyLogPopUpCommand;
+        private ICommand openAddLogPopUpCommand;
+        private ICommand closeAddLogPopUpCommand;
+        private ICommand createPdfCommand;
 
 
         public ICommand PopUpAddCommand => popUpAddCommand ??= new RelayCommand(PopUpAdd);
@@ -39,6 +55,17 @@ namespace SWE2TourPlanner.ViewModels
         public ICommand PopUpModifyLogCommand => popUpModifyLogCommand ??= new RelayCommand(PopUpModifyLog);
         public ICommand SearchCommand => searchCommand ??= new RelayCommand(Search);
         public ICommand ClearCommand => clearCommand ??= new RelayCommand(Clear);
+        public ICommand OpenDeleteTourPopUpCommand => openDeleteTourPopUpCommand ??= new RelayCommand(OpenDeleteTourPopUp);
+        public ICommand CloseDeleteTourPopUpCommand => closeDeleteTourPopUpCommand ??= new RelayCommand(CloseDeleteTourPopUp);
+        public ICommand OpenModifyTourPopUpCommand => openModifyTourPopUpCommand ??= new RelayCommand(OpenModifyTourPopUp);
+        public ICommand CloseModifyTourPopUpCommand => closeModifyTourPopUpCommand ??= new RelayCommand(CloseModifyTourPopUp);
+        public ICommand OpenDeleteLogPopUpCommand => openDeleteLogPopUpCommand ??= new RelayCommand(OpenDeleteLogPopUp);
+        public ICommand CloseDeleteLogPopUpCommand => closeDeleteLogPopUpCommand ??= new RelayCommand(CloseDeleteLogPopUp);
+        public ICommand OpenModifyLogPopUpCommand => openModifyLogPopUpCommand ??= new RelayCommand(OpenModifyLogPopUp);
+        public ICommand CloseModifyLogPopUpCommand => closeModifyLogPopUpCommand ??= new RelayCommand(CloseModifyLogPopUp);
+        public ICommand OpenAddLogPopUpCommand => openAddLogPopUpCommand ??= new RelayCommand(OpenAddLogPopUp);
+        public ICommand CloseAddLogPopUpCommand => closeAddLogPopUpCommand ??= new RelayCommand(CloseAddLogPopUp);
+        public ICommand CreatePdfCommand => createPdfCommand ??= new RelayCommand(CreatePdf);
 
         //public TourAddViewModel tourAddViewModel;
         public LogAddViewModel logAddViewModel;
@@ -82,6 +109,71 @@ namespace SWE2TourPlanner.ViewModels
             }
         }
 
+        public bool IsPopUpAddLogVisible
+        {
+            get { return isPopUpAddLogVisible; }
+            set
+            {
+                if (isPopUpAddLogVisible != value)
+                {
+                    isPopUpAddLogVisible = value;
+                    RaisePropertyChangedEvent(nameof(IsPopUpAddLogVisible));
+                }
+            }
+        }
+
+        public bool IsPopUpDeleteTourVisible
+        {
+            get { return isPopUpDeleteTourVisible; }
+            set
+            {
+                if(isPopUpDeleteTourVisible != value)
+                {
+                    isPopUpDeleteTourVisible = value;
+                    RaisePropertyChangedEvent(nameof(IsPopUpDeleteTourVisible));
+                }
+            }
+        }
+
+        public bool IsPopUpModifyTourVisible
+        {
+            get { return isPopUpModifyTourVisible; }
+            set
+            {
+                if(isPopUpModifyTourVisible != value)
+                {
+                    isPopUpModifyTourVisible = value;
+                    RaisePropertyChangedEvent(nameof(IsPopUpModifyTourVisible));
+                }
+            }
+        }
+
+        public bool IsPopUpDeleteLogVisible
+        {
+            get { return isPopUpDeleteLogVisible; }
+            set
+            {
+                if (isPopUpDeleteLogVisible != value)
+                {
+                    isPopUpDeleteLogVisible = value;
+                    RaisePropertyChangedEvent(nameof(IsPopUpDeleteLogVisible));
+                }
+            }
+        }
+
+        public bool IsPopUpModifyLogVisible
+        {
+            get { return isPopUpModifyLogVisible; }
+            set
+            {
+                if (isPopUpModifyLogVisible != value)
+                {
+                    isPopUpModifyLogVisible = value;
+                    RaisePropertyChangedEvent(nameof(IsPopUpModifyLogVisible));
+                }
+            }
+        }
+
         public MainViewModel()
         {
             this.tourItemFactory = TourItemFactory.GetInstance();
@@ -119,18 +211,21 @@ namespace SWE2TourPlanner.ViewModels
 
         private void PopUpAdd(object commandParameter)
         {
-            //this.tourAddViewModel = new TourAddViewModel();
-            TourAddWindow view = new TourAddWindow();
-            view.DataContext = new TourAddViewModel();
-            view.ShowDialog();
+            if(!IsPopUpDeleteTourVisible && !IsPopUpModifyTourVisible && !IsPopUpDeleteLogVisible && !IsPopUpModifyTourVisible)
+            {
+                //this.tourAddViewModel = new TourAddViewModel();
+                TourAddWindow view = new TourAddWindow();
+                view.DataContext = new TourAddViewModel();
+                view.ShowDialog();
 
-            TourItems.Clear();
-            FillListBox();
+                TourItems.Clear();
+                FillListBox();
+            }
         }
 
         private void PopUpAddLog(object commandParameter)
         {
-            if(CurrentItem != null)
+            if(CurrentItem != null && !IsPopUpAddLogVisible && !IsPopUpDeleteTourVisible && !IsPopUpModifyTourVisible && !IsPopUpDeleteLogVisible && !IsPopUpModifyTourVisible)
             {
                 this.logAddViewModel = new LogAddViewModel();
                 logAddViewModel.CurrentTour = CurrentItem;
@@ -142,27 +237,45 @@ namespace SWE2TourPlanner.ViewModels
                 LogItems.Clear();
                 FillDataGrid(CurrentItem.TourId);
             }
+            else if(CurrentItem == null && !IsPopUpAddLogVisible && !IsPopUpDeleteTourVisible && !IsPopUpModifyTourVisible && !IsPopUpDeleteLogVisible && !IsPopUpModifyLogVisible)
+            {
+                OpenAddLogPopUp(commandParameter);
+            }
         }
 
         private void DeleteTour(object commandParameter)
         {
-            string path = CurrentItem.ImagePath;
-            this.tourItemFactory.DeleteItemAndSavePath(CurrentItem.TourId, path);
-            TourItems.Clear();
-            FillListBox();
-            LogItems.Clear();
+            if(CurrentItem != null && !IsPopUpDeleteTourVisible && !IsPopUpModifyTourVisible && !IsPopUpDeleteLogVisible && !IsPopUpModifyLogVisible)
+            {
+                string path = CurrentItem.ImagePath;
+                this.tourItemFactory.DeleteItemAndSavePath(CurrentItem.TourId, path);
+                TourItems.Clear();
+                FillListBox();
+                LogItems.Clear();
+            }
+            else if(CurrentItem == null && !IsPopUpDeleteTourVisible && !IsPopUpModifyTourVisible && !IsPopUpDeleteLogVisible && !IsPopUpModifyLogVisible)
+            {
+                OpenDeleteTourPopUp(commandParameter);
+            }
         }
-
+        
         private void DeleteLog(object commandParameter)
         {
-            this.tourItemFactory.DeleteLog(CurrentLog.LogId);
-            LogItems.Clear();
-            FillDataGrid(CurrentItem.TourId);
+            if (CurrentLog != null && !IsPopUpDeleteTourVisible && !IsPopUpModifyTourVisible && !IsPopUpDeleteLogVisible && !IsPopUpModifyLogVisible)
+            {
+                this.tourItemFactory.DeleteLog(CurrentLog.LogId);
+                LogItems.Clear();
+                FillDataGrid(CurrentItem.TourId);
+            }
+            else if (CurrentLog == null && !IsPopUpDeleteTourVisible && !IsPopUpModifyTourVisible && !IsPopUpDeleteLogVisible && !IsPopUpModifyLogVisible)
+            {
+                OpenDeleteLogPopUp(commandParameter);
+            }
         }
 
         private void PopUpModify(object commandParameter)
         {
-            if(CurrentItem != null)
+            if (CurrentItem != null && !IsPopUpDeleteTourVisible && !IsPopUpModifyTourVisible && !IsPopUpDeleteLogVisible && !IsPopUpModifyLogVisible)
             {
                 this.tourModifyViewModel = new TourModifyViewModel();
                 tourModifyViewModel.CurrentTour = CurrentItem;
@@ -174,11 +287,15 @@ namespace SWE2TourPlanner.ViewModels
                 TourItems.Clear();
                 FillListBox();
             }
+            else if (CurrentItem == null && !IsPopUpDeleteTourVisible && !IsPopUpModifyTourVisible && !IsPopUpDeleteLogVisible && !IsPopUpModifyLogVisible)
+            {
+                OpenModifyTourPopUp(commandParameter);
+            }
         }
 
         private void PopUpModifyLog(object commandParameter)
         {
-            if (CurrentLog != null)
+            if (CurrentLog != null && !IsPopUpDeleteTourVisible && !IsPopUpModifyTourVisible && !IsPopUpDeleteLogVisible && !IsPopUpModifyLogVisible)
             {
                 this.logModifyViewModel = new LogModifyViewModel();
                 logModifyViewModel.CurrentLog = CurrentLog;
@@ -189,6 +306,10 @@ namespace SWE2TourPlanner.ViewModels
 
                 LogItems.Clear();
                 FillDataGrid(CurrentItem.TourId);
+            }
+            else if (CurrentLog == null && !IsPopUpDeleteTourVisible && !IsPopUpModifyTourVisible && !IsPopUpDeleteLogVisible && !IsPopUpModifyLogVisible)
+            {
+                OpenModifyLogPopUp(commandParameter);
             }
         }
 
@@ -220,6 +341,61 @@ namespace SWE2TourPlanner.ViewModels
             TourItems.Clear();
             SearchText = "";
             FillListBox();
+        }
+
+        private void CreatePdf(object commandParameter)
+        {
+            tourItemFactory.CreatePdf(CurrentItem);
+        }
+
+        private void OpenDeleteTourPopUp(object commandParameter)
+        {
+            IsPopUpDeleteTourVisible = true;
+        }
+
+        private void CloseDeleteTourPopUp(object commandParameter)
+        {
+            IsPopUpDeleteTourVisible = false;
+        }
+
+        private void OpenModifyTourPopUp(object commandParameter)
+        {
+            IsPopUpModifyTourVisible = true;
+        }
+
+        private void CloseModifyTourPopUp(object commandParameter)
+        {
+            IsPopUpModifyTourVisible = false;
+        }
+
+        private void OpenDeleteLogPopUp(object commandParameter)
+        {
+            IsPopUpDeleteLogVisible = true;
+        }
+
+        private void CloseDeleteLogPopUp(object commandParameter)
+        {
+            IsPopUpDeleteLogVisible = false;
+        }
+
+        private void OpenModifyLogPopUp(object commandParameter)
+        {
+            IsPopUpModifyLogVisible = true;
+        }
+
+        private void CloseModifyLogPopUp(object commandParameter)
+        {
+            IsPopUpModifyLogVisible = false;
+        }
+
+        private void OpenAddLogPopUp(object commandParameter)
+        {
+            IsPopUpAddLogVisible = true;
+        }
+
+        private void CloseAddLogPopUp(object commandParameter)
+        {
+            IsPopUpAddLogVisible = false;
         }
     }
 }
